feat(dashboard): draw the real dive profile chart on the home tab
The latest-dive slot on the home tab drew a bespoke LineChart: depth only,
no legend, no analysis curves, no markers, no gestures. It looked nothing
like the chart the same dive gets on the dive detail page.

Extract the detail page's chart wiring into DiveProfileChartHost, a single
widget that turns a hydrated Dive into a fully configured DiveProfileChart:
per-source profiles and their analyses, source overlays, the planned-profile
ghost, max-depth and pressure-threshold markers, photo pins, estimated tank
pressures, gas segments and switches, the safety-findings lane, range
selection, playback and cursor tracking. Only host-specific concerns stay
parameters. The detail page now renders through it, so it stays the reference
and the two surfaces cannot drift apart.

profileLegendProvider is a single app-wide notifier rather than a per-dive
family, so every metric toggle, right-axis choice and per-tank visibility the
diver sets in dive details already applies on the home tab.

The home card re-reads the newest dive through diveProvider instead of
charting the recent-dives list copy: that list refreshes on dives-table
writes only, so a reparse rewriting samples in place would have left a stale
shape on screen. The whole-card InkWell is gone, since it would swallow the
chart's own pan, zoom, tooltip and range gestures; the header is the tap
target and gains a fullscreen button. The slot grows to 380px to seat the
legend row above the plot, which still leaves the plot more height than the
old static preview had.

Drop latestDiveProfileProvider. It only fed this preview, and it re-fetched
samples the dive already carried: its comment claimed getDiveById leaves
Dive.profile empty, but that call goes through _mapRowToDive, which hydrates
tanks, tank series and the merged profile.
