+++ lib/features/dashboard/presentation/providers/dashboard_providers.dart
@@ -51,36 +51,6 @@ final recentDivesProvider = FutureProvider<List<Dive>>((ref) async {
   return recent;
 });
 
-/// Depth profile of the newest dive, for the recent-dives preview chart.
-///
-/// [recentDivesProvider] hydrates its dives through `getDiveById`, which reads
-/// the `dives` row only and leaves `Dive.profile` empty -- the list path
-/// deliberately avoids dragging thousands of sample rows into a summary view.
-/// The batch cache it does populate holds ~20-point sparkline summaries, which
-/// is enough for a tile-sized mini chart but too coarse for a preview several
-/// hundred pixels wide. So this fetches the real samples, for the newest dive
-/// only, and only when a caller actually watches it.
-///
-/// Null when there are no dives, or when the newest dive has no profile
-/// (manually logged dives usually do not).
-final latestDiveProfileProvider = FutureProvider<List<DiveProfilePoint>?>((
-  ref,
-) async {
-  final repository = ref.watch(diveRepositoryProvider);
-  // The dive-detail tick, not the dives tick: this reads
-  // dive_profile_series, and
-  // samples change without the dives row changing (a reparse or a sync pull
-  // rewrites the profile in place). watchDivesChanges would leave the chart
-  // showing the pre-reparse shape.
-  ref.invalidateSelfWhen(repository.watchDiveDetailChanges());
-
-  final recent = await ref.watch(recentDivesProvider.future);
-  if (recent.isEmpty) return null;
-
-  final profile = await repository.getDiveProfile(recent.first.id);
-  return profile.isEmpty ? null : profile;
-});
-
 /// Current diver provider (re-exported for convenience)
 final dashboardDiverProvider = FutureProvider<Diver?>((ref) async {
   return ref.watch(currentDiverProvider.future);
